chore: fix release publishing workflow
Instead of me having to manually push a git tag, release-please takes
care of that.
The previous pipeline expected a tag to be pushed which kicks off the
release process.

Instead, the release process should start once release-please creates a
draft release.

This should prevent the case where I have to manually create a release
and do it incorrectly.

Fixes: https://github.com/OGKevin/obsidian-kobo-highlights-import/issues/143
Release-As: 1.5.2